backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
openai==0.28.1
PyPDF2==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0